requirements.txt
aioredis==1.3.1
alabaster==0.7.12
anyio==3.3.4
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
black==21.10b0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
clikit==0.6.2
commonmark==0.9.1
crashtest==0.3.1
distlib==0.3.3
docutils==0.17.1
fastapi==0.70.0
fastapi-cache==0.1.0
fastapi-cache2==0.1.7
filelock==3.3.2
h11==0.12.0
hiredis==2.0.0
html5lib==1.1
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
install==1.3.4
Jinja2==3.0.2
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
pastel==0.2.1
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
poetry==1.1.11
poetry-core==1.0.7
ptyprocess==0.7.0
py==1.11.0
pydantic==1.8.2
Pygments==2.10.0
pylev==1.4.0
pyparsing==3.0.4
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytzdata==2020.1
recommonmark==0.7.1
regex==2021.11.2
requests==2.26.0
requests-toolbelt==0.9.1
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.3
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.7.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
webencodings==0.5.1
